Fix diamond pattern, hard mode, and bullet spawner behavior

See merge request cpts487-sp2021/team-madison!82

+++ TestProject/core/src/Objects/GameObject/Entity.java
@@ -165,7 +165,7 @@ public abstract class Entity extends GameObject {
         return this.speed;
     }
 
-    public void setSpeed(int speed) {
+    public void setSpeed(float speed) {
         this.speed = speed;
     }
